+++ tests/testthat/_snaps/auto/NEWS.md
@@ -4,7 +4,7 @@
 
 - Add cool blop.
 
-- Add cool bla.
+- Add cool bla (@someone, #hashtag-test).
 
 - Added a `NEWS.md` file to track changes to the package.